--- ignition/parameters/FxSave.json
@@ -1,19 +1,0 @@
-{
-  "ProxyAdmin": {
-    "Fx": "0x9B54B7703551D9d0ced177A78367560a8B2eDDA4"
-  },
-  "EmptyContract": {
-    "deployed": "0x387568e1ea4Ff4D003B8147739dB69D87325E206"
-  },
-  "FxSave": {
-    "fxBase": "0x65C9A641afCEB9C0E6034e558A319488FA0FA3be",
-    "fxBaseGauge": "0xEd92dDe3214c24Ae04F5f96927E3bE8f8DbC3289",
-    "LiquidityGaugeImplementation": "0xF62F458D2F6dd2AD074E715655064d7632e136D6",
-    "Name": "f(x) USD Saving",
-    "Symbol": "fxSAVE",
-    "pid": 36,
-    "Threshold": "100000000000000000000",
-    "Treasury": "0x0084C2e1B1823564e597Ff4848a88D61ac63D703",
-    "Router": "0x33636D49FbefBE798e15e7F356E8DBef543CC708"
-  }
-}
