+++ ignition/parameters/FxSaveAndWBTCPool.json
@@ -1,0 +1,41 @@
+{
+  "ProxyAdmin": {
+    "Fx": "0x9B54B7703551D9d0ced177A78367560a8B2eDDA4",
+    "Custom": "0xd41d29fc53fE5Ce9f0fB2328E54d35A2a03a324B"
+  },
+  "EmptyContract": {
+    "deployed": "0x387568e1ea4Ff4D003B8147739dB69D87325E206"
+  },
+  "Upgrade20250318": {
+    "Router": "0x33636D49FbefBE798e15e7F356E8DBef543CC708",
+    "FxUSDProxy": "0x085780639CC2cACd35E474e71f4d000e2405d8f6",
+    "FxUSDBasePoolProxy": "0x65C9A641afCEB9C0E6034e558A319488FA0FA3be",
+    "PegKeeperProxy": "0x50562fe7e870420F5AAe480B7F94EB4ace2fcd70",
+    "PoolManagerProxy": "0x250893CA4Ba5d05626C785e8da758026928FCD24",
+    "LendingPool": "0x87870Bca3F3fD6335C3F4ce8392D69350B4fA4E2",
+    "BaseAsset": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
+  },
+  "FxSaveAndWBTCPool": {
+    "FxUSDBasePoolGaugeProxyV1": "0xEd92dDe3214c24Ae04F5f96927E3bE8f8DbC3289",
+    "LiquidityGaugeImplementation": "0xF62F458D2F6dd2AD074E715655064d7632e136D6",
+    "FxSaveName": "f(x) USD Saving",
+    "FxSaveSymbol": "fxSAVE",
+    "pid": 36,
+    "Threshold": "100000000000000000000",
+    "Treasury": "0x0084C2e1B1823564e597Ff4848a88D61ac63D703",
+
+    "SpotPriceOracle": "0xc2312CaF0De62eC9b4ADC785C79851Cb989C9abc",
+    "FxWBTCName": "f(x) WBTC position",
+    "FxWBTCSymbol": "xWBTC",
+    "DebtRatioLower": "90909090909090909",
+    "DebtRatioUpper": "866666666666666666",
+    "RebalanceDebtRatio": "880000000000000000",
+    "RebalanceBonusRatio": "25000000",
+    "LiquidateDebtRatio": "950000000000000000",
+    "LiquidateBonusRatio": "40000000",
+    "OpenRatio": "3000000",
+    "OpenRatioStep": "300000000000000000",
+    "CloseFeeRatio": "1000000",
+    "FundingRatio": "1000000000"
+  }
+}
