--- ignition/parameters/multiple_collateral.json
@@ -1,47 +1,0 @@
-{
-  "Upgrade202502xx": {
-    "Router": "0x33636D49FbefBE798e15e7F356E8DBef543CC708",
-    "FxUSDProxy": "0x085780639CC2cACd35E474e71f4d000e2405d8f6",
-    "FxUSDBasePoolProxy": "0x65C9A641afCEB9C0E6034e558A319488FA0FA3be",
-    "PegKeeperProxy": "0x50562fe7e870420F5AAe480B7F94EB4ace2fcd70",
-    "PoolManagerProxy": "0x250893CA4Ba5d05626C785e8da758026928FCD24",
-    "LendingPool": "0x87870Bca3F3fD6335C3F4ce8392D69350B4fA4E2",
-    "BaseAsset": "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
-  },
-  "WETHPool": {
-    "ProxyAdmin": "0x9B54B7703551D9d0ced177A78367560a8B2eDDA4",
-    "SpotPriceOracle": "0xc2312CaF0De62eC9b4ADC785C79851Cb989C9abc",
-    "Name": "f(x) WETH position",
-    "Symbol": "xWETH",
-    "CollateralCapacity": "50000000000000000000000",
-    "DebtCapacity": "22500000000000000000000000",
-    "DebtRatioLower": "90909090909090909",
-    "DebtRatioUpper": "866666666666666666",
-    "RebalanceDebtRatio": "880000000000000000",
-    "RebalanceBonusRatio": "25000000",
-    "LiquidateDebtRatio": "950000000000000000",
-    "LiquidateBonusRatio": "40000000",
-    "OpenRatio": "1000000",
-    "OpenRatioStep": "300000000000000000",
-    "CloseFeeRatio": "1000000",
-    "FundingRatio": "1000000000"
-  },
-  "WBTCPool": {
-    "ProxyAdmin": "0x9B54B7703551D9d0ced177A78367560a8B2eDDA4",
-    "SpotPriceOracle": "0xc2312CaF0De62eC9b4ADC785C79851Cb989C9abc",
-    "Name": "f(x) WBTC position",
-    "Symbol": "xWBTC",
-    "CollateralCapacity": "100000000000",
-    "DebtCapacity": "22500000000000000000000000",
-    "DebtRatioLower": "90909090909090909",
-    "DebtRatioUpper": "866666666666666666",
-    "RebalanceDebtRatio": "880000000000000000",
-    "RebalanceBonusRatio": "25000000",
-    "LiquidateDebtRatio": "950000000000000000",
-    "LiquidateBonusRatio": "40000000",
-    "OpenRatio": "1000000",
-    "OpenRatioStep": "300000000000000000",
-    "CloseFeeRatio": "1000000",
-    "FundingRatio": "1000000000"
-  }
-}
